Use lexical binding
This package does not appear to depend on dynamic binding. Lexical
binding has been a thing since Emacs 24. "Special" variables (those
defined with defvar and similar) are still dynamic in lexical binding
mode.

In general, lexical binding is either a bit faster or has zero impact,
and should always be turned on. Details:
https://nullprogram.com/blog/2016/12/22/